requirements.txt
aioredis==1.3.1
asgi-redis==1.4.3   
asgiref==3.2.7      
astroid==2.3.3      
async-timeout==3.0.1
attrs==19.3.0       
autobahn==20.4.3    
Automat==20.2.0     
certifi==2020.4.5.1 
cffi==1.14.0        
chardet==3.0.4      
colorama==0.4.3
constantly==15.1.0
cryptography==2.9.2
dj-database-url==0.5.0
Django==3.0.6
django-appconf==1.0.4
django-ckeditor==5.9.0
django-classy-tags==1.0.0
django-cms==3.7.1
django-crispy-forms==1.9.0
django-entangled==0.3
django-filer==1.7.0
django-formtools==2.2
django-imagekit==4.0.2
django-js-asset==1.2.2
django-mptt==0.11.0
django-online-users==0.3
django-polymorphic==2.1.2
django-search-views==0.3.7
django-sekizai==1.1.0
django-taggit==1.2.0
django-treebeard==4.3.1
djangocms-admin-style==1.4.0
djangocms-attributes-field==1.2.0        
djangocms-bootstrap4==1.5.0
djangocms-cascade==1.2.3
djangocms-file==2.4.0
djangocms-googlemap==1.3.0
djangocms-icon==1.5.0
djangocms-installer==1.2.1
djangocms-link==2.5.0
djangocms-picture==2.3.0
djangocms-snippet==2.2.0
djangocms-style==2.2.0
djangocms-text-ckeditor==3.8.0
djangocms-video==2.2.0
easy-thumbnails==2.7
gunicorn==20.0.4
hiredis==1.0.1
html5lib==1.0.1
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
isort==4.3.21
jsonfield==3.1.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
msgpack==0.6.2
msgpack-python==0.5.6
pilkit==2.0
Pillow==7.1.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pycparser==2.20
PyHamcrest==2.0.2
pylint==2.4.4
pyOpenSSL==19.1.0
pytz==2019.3
redis==2.10.6
requests==2.23.0
service-identity==18.1.0
six==1.14.0
sqlparse==0.3.1
txaio==20.4.1
tzlocal==2.0.0
Unidecode==1.1.1
urllib3==1.25.9
webencodings==0.5.1
whitenoise==5.1.0
wrapt==1.11.2
zope.interface==5.1.0